Implement deletion of singular contact + its tests, update documentation

+++ src/repositories/contact_repository.py
@@ -76,16 +76,17 @@ class ContactRepository:
         cursor.execute("DELETE FROM contacts WHERE user = ?", (user_id, ))
         self._connection.commit()
 
-    def delete_one(self, contact_id):
+    def delete_one(self, user_id, contact_id):
         """Poistaa yhden kontaktin annetun id'n perusteella.
 
         Args:
+            user_id: käyttäjän id, jonka luoma kontakti poistetaan
             contact_id: poistettavan kontaktin id 
         """
 
         cursor = self._connection.cursor()
         cursor.execute(
-            "DELETE FROM contacts WHERE id = ?", (contact_id, ))
+            "DELETE FROM contacts WHERE id = ? AND user = ?", (contact_id, user_id, ))
         self._connection.commit()
 
     def create(self, contact):
